# Optimize ctl call and precompute gap shift