Some more optimizations that worked well for me, building on top of #1.

```
previous from PR1
k==15, time: 1.284619s, enc/s: 155688129
k==16, time: 1.063841s, enc/s: 187997971

more optimizations
k==15, time: 1.073328s, enc/s: 186336348
k==16, time: 1.029689s, enc/s: 194233362
```

In my C++ version of this code, see [here](https://github.com/lczech/genesis/blob/dev/lib/genesis/sequence/kmer/canonical_encoding.hpp), I'm achieving even higher throughput (>200mio enc/s). Not quite sure why the C equivalent is a bit behind here. But well, good enough; I think for now, I'll leave it at that.